+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- endpoint to get a list of traversing paths from datasets to the parent folders of the given file
+  [CATS-811](https://opensource.ncsa.illinois.edu/jira/browse/CATS-811)
 - Added clowder.upload.previews flag to application.conf to turn on/off previews in upload page.
   [CATS-813](https://opensource.ncsa.illinois.edu/jira/browse/CATS-813)